+++ examples/resources/familio_event/import.sh
@@ -1,0 +1,3 @@
+# familio_event is imported by "<person_uuid>:<event_uuid>" (an event is a
+# sub-resource of a person and has no global address).
+terraform import familio_event.ivan_residence 121681c3-1234-5678-9abc-def012345678:ed0a313b-8765-4321-fedc-ba9876543210